medaide_plus_trial/requirements.txt
transformers>=4.40.0
torch>=2.0.0
sentence-transformers>=2.7.0
rank-bm25>=0.2.2
networkx>=3.3
faiss-cpu>=1.7.4
scikit-learn>=1.4.0
numpy>=1.26.0
scipy>=1.12.0
spacy>=3.7.0
openai>=1.30.0
langchain>=0.2.0
langchain-openai>=0.1.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
gradio>=4.37.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
tqdm>=4.66.0
matplotlib>=3.8.0
seaborn>=0.13.0
# Phase 4 evaluation dependencies
nltk>=3.8.1
rouge-score>=0.1.2
bert-score>=0.3.13
datasets>=2.19.0
# Multi-provider LLM support
httpx>=0.27.0
anthropic>=0.25.0
google-generativeai>=0.5.0
